+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Permission and idle notification phrases now vary randomly instead of repeating the same line
+
 ## [0.2.0] - 2026-02-25
 
 ### Added